refactor(s3stream): refactor ObjectStorageTest
